removed methods that invokes keycloakHelper.js method in notification.js
